Rename to database-returns
