+++ src/lib/types/index.ts
@@ -1,4 +1,4 @@
-export type * from './action-returns';
 export type * from './api-responses';
+export type * from './database-returns';
 export type * from './hooks';
 export type * from './schemas';
